afrs-microservice/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
pydantic==1.10.11
mysql-connector-python==8.1.0
python-multipart==0.0.6
cryptography==41.0.3
opencv-python==4.8.0.76
ultralytics==8.0.223
transformers==4.42.0
torch==2.2.2
faiss-cpu==1.7.4
Pillow==10.0.0
huggingface-hub==0.23.4
PyJWT==2.8.0
pika==1.3.2
httpx
pytest